+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## v0.0.26 [unreleased]
+## v0.1.0 [unreleased]
 
 ## Features
 
 - Add documentation for all template parameters (`awless create instance -h`, `awless update s3object -h`...)
 - Listing with filter invalid keys: return error and help
+- `awless whoami` now has flags to return specific account properties only: `--account-only`, `--id-only`, `--name-only`, `--resource-only`, `--type-only`
 - Rename template parameters for standardization:
     - `delete keypair id=...` -> `delete keypair name=...`
     - `create listener target=...` -> `create listener targetgroup=...`